fix more various things, unify table for functions, variables, events, etc
